requirements.txt
discord.py
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib